refactor(prisma): move PrismaModule & PrismaService into src/prisma, fix imports and adjust deploy configs
