+++ server/.env.example
@@ -1,3 +1,4 @@
+PORT
 DATABASE_URL
 JWT_SECRET
 JWT_EXPIRES_IN
