+++ server/src/auth/auth.service.ts
@@ -9,7 +9,7 @@ import { JwtService } from '@nestjs/jwt';
 import { RegisterRequestDto } from './dto/register.request.dto';
 import { AuthResponseDto } from './dto/auth.response.dto';
 import { UserResponseDto } from './dto/auth.response.dto';
-import { PrismaService } from '../../prisma/prisma.service';
+import { PrismaService } from '../prisma/prisma.service';
 import * as bcrypt from 'bcrypt';
 
 @Injectable()
